Refactor setting language
