+++ src-capacitor/android/app/src/main/assets/capacitor.config.json
@@ -13,6 +13,6 @@
 	},
 	"server": {
 		"androidScheme": "http",
-		"url": "http://192.168.0.200:9500"
+		"url": "http://192.168.0.20:9500"
 	}
 }
