+++ src/scripts/staticData/languages.ts
@@ -1,4 +1,4 @@
-import { Language, SelectOption } from '../models'
+import { Language } from '../models'
 
 export function getLanguages(): Language[] {
   return [
@@ -6,14 +6,4 @@ export function getLanguages(): Language[] {
     { id: 2, text: 'English', code: 'en' },
     { id: 3, text: 'Deutsch', code: 'de' }
   ]
-}
-
-export function getLanguageOptions(): SelectOption[] {
-  return getLanguages().map(
-    l =>
-      ({
-        label: l.text,
-        value: l.id
-      }) as SelectOption
-  )
 }
